icon replacement in Decks list and edit card window; Default deck implementation; tag colour scheme change

+++ src/renderer/src/components/decks/DeckPage.tsx
@@ -1,8 +1,10 @@
 import { useEffect, useState } from 'react'
 import { Deck } from 'src/types/types'
+import DeleteIcon from '../cards/DeleteIcon'
+import AddSignIcon from '../cards/AddSignIcon'
+const DEFAULT_DECK_NAME: string = 'DEFAULT'
 
 function AddDeckForm({ decks }: { decks: Deck[] }): JSX.Element | string {
-
   function addDeck(): void {
     const deckNameForm = document.getElementById('deckNameInput') as HTMLInputElement
     const deckName = deckNameForm.value.toUpperCase()
@@ -12,24 +14,23 @@ function AddDeckForm({ decks }: { decks: Deck[] }): JSX.Element | string {
       cards: []
     }
 
-    if(decks.find(x => x.name.localeCompare(deckName) == 0)) {
-      triggerToast(false, "Name already exists!")
-    }
-    else if(!deckName.trim()) {
-      triggerToast(false, "Must enter text!")
-    }
-    else {
-      deckNameForm.value = "";
-      window.api.store.addDeck(newDeck);
-      triggerToast(true, "Added Deck!")
+    if (decks.find((x) => x.name.localeCompare(deckName) == 0)) {
+      triggerToast(false, 'Name already exists!')
+    } else if (!deckName.trim()) {
+      triggerToast(false, 'Must enter text!')
+    } else {
+      deckNameForm.value = ''
+      window.api.store.addDeck(newDeck)
+      triggerToast(true, 'Added Deck!')
     }
   }
 
   return (
     <div className="pt-2">
-      <input id="deckNameInput" placeholder="Enter new deck...">
-      </input>
-      <button onClick={addDeck}>Add</button>
+      <input id="deckNameInput" placeholder="Enter new deck..."></input>
+      <button onClick={addDeck} className="align-middle p-5">
+        <AddSignIcon width="30px" height="30px" />
+      </button>
     </div>
   )
 }
@@ -42,8 +43,10 @@ function DeckTable({
   toCardsByDeckBtn: Function
 }): JSX.Element | string {
   function deleteDeck(name: string): void {
-    triggerToast(true, "Deleted Deck!")
-    window.api.store.deleteDeck(name);
+    if (name !== DEFAULT_DECK_NAME) {
+      triggerToast(true, 'Deleted Deck!')
+      window.api.store.deleteDeck(name)
+    } else alert(DEFAULT_DECK_NAME + ' cannot be deleted!')
   }
   const deckEntries = decks.map((deck) => (
     <tr className="bg-slate-800 border" height="50px" key={deck.name}>
@@ -52,8 +55,8 @@ function DeckTable({
       </td>
 
       <td>
-        <button className="pl-10" onClick={() => deleteDeck(deck.name)}>
-          Delete
+        <button className="pl-10 align-middle" onClick={() => deleteDeck(deck.name)}>
+          <DeleteIcon width="30px" height="30px" />
         </button>
       </td>
     </tr>
@@ -70,25 +73,22 @@ function DeckTable({
   )
 }
 
-function triggerToast(positive: boolean, text: string){
-
-  var timeVisible = 3;
-  var uptime = setInterval(function(){
-
-    if(timeVisible <= 0) {
-      (document.getElementById("deckToast")! as HTMLElement).className =  "invisible";
-      clearInterval(uptime);
+function triggerToast(positive: boolean, text: string) {
+  let timeVisible = 3
+  var uptime = setInterval(function () {
+    if (timeVisible <= 0) {
+      ;(document.getElementById('deckToast')! as HTMLElement).className = 'invisible'
+      clearInterval(uptime)
+    } else {
+      ;(document.getElementById('deckToast')! as HTMLElement).className = positive
+        ? 'rounded p-2 outline outline-green-500 bg-white absolute top-10 right-12'
+        : 'rounded p-2 outline outline-red-500 bg-white absolute top-10 right-12'
+      ;(document.getElementById('deckToast')! as HTMLElement).innerHTML =
+        text + ' ' + timeVisible.toString()
+      timeVisible -= 1
     }
-    else {
-      (document.getElementById("deckToast")! as HTMLElement).className = positive ? 
-        "rounded p-2 outline outline-green-500 bg-white absolute top-10 right-12" : "rounded p-2 outline outline-red-500 bg-white absolute top-10 right-12";
-
-      (document.getElementById("deckToast")! as HTMLElement).innerHTML = text + " " + timeVisible.toString();
-      timeVisible -= 1;
-    }
-  }, 1000);
+  }, 1000)
 }
-
 
 function DecksPage({ toCardsByDeckBtn }: { toCardsByDeckBtn: Function }): JSX.Element {
   const [decks, setDecks] = useState<Deck[]>([])
@@ -107,7 +107,7 @@ function DecksPage({ toCardsByDeckBtn }: { toCardsByDeckBtn: Function }): JSX.El
 
   return (
     <div className="pt-5">
-      <DeckTable decks={decks} toCardsByDeckBtn={toCardsByDeckBtn}/>
+      <DeckTable decks={decks} toCardsByDeckBtn={toCardsByDeckBtn} />
       <AddDeckForm decks={decks} />
       <div id="deckToast" className="invisible"></div>
     </div>
@@ -115,3 +115,4 @@ function DecksPage({ toCardsByDeckBtn }: { toCardsByDeckBtn: Function }): JSX.El
 }
 
 export default DecksPage
+export { DEFAULT_DECK_NAME }
